[C8] Split part.go by translation stage (#2231)
model/exchange/translators/inventor/translate/part.go was 1806 LOC, 3.6x the
500-line ceiling — the largest file in the milestone. Split the Inventor .ipt
part translator by translation stage; each file is now under the ceiling:

- part.go (354): entry points + the feature-graph walk (BodyFromIPT,
  FromInventor, buildPart, addFeatures), the solidity/decline diagnostics
  (firstBodyIsSolid/nearlySolid/openEdgeCount/retryWithoutDissolve) and the
  display-mesh body import.
- part_params.go (106): parameter/expression mapping (addParameters,
  operationOf, extentOf, directionOf).
- part_sketch.go (254): sketch translation (extractSketches, plane resolution,
  emitSketch/emitSketchOn and the emit helpers).
- part_revolve.go (252): the revolve feature builder (detection + build).
- part_features.go (262): the other per-feature-kind builders — extrude,
  mirror, rectangular/circular pattern, hole, loft, sweep.
- part_constraints_geometric.go (269): geometric constraint application
  (geometric/tangent/circle-relation/ground/symmetry, applyGeoConstraint).
- part_constraints_dimensional.go (409): dimensional constraint application
  (distance/angle/offset/radius/revolve-radius/axial/centreline) + the shared
  sketch-entity coordinate lookups.

Behaviour-preserving: split programmatically by top-level declaration, then
verified every non-comment code line is byte-identical to the original (only
per-file import redistribution and gofmt whitespace differ); imports resolved
per file with goimports. The translate corpus passes before and after
(GOWORK=off, separate module); whole inventor module green; translate lint
issue count unchanged (11, all pre-existing). Net delta of recognizers /
tolerance constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2231.

(cherry picked from commit 68195285fef9eaf510bc2990b0bd0999e444e8fc)
